pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "libro"
version = "0.1.0"
description = "KDP niche detection and book creation system"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "sqlalchemy>=2.0",
    "alembic>=1.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "playwright>=1.40",
    "httpx>=0.25",
    "pillow>=10.0",
    "reportlab>=4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
libro = "libro.cli:app"

[tool.setuptools.packages.find]
include = ["libro*"]
